+++ app/src/main/java/com/tfandkusu/ate/feature/jnta/Screen1Fragment.kt
@@ -4,8 +4,11 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.compose.foundation.layout.Arrangement.Absolute.spacedBy
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -21,12 +24,7 @@ import com.tfandkusu.ate.R
 import com.tfandkusu.ate.component.MyTopAppBar
 import com.tfandkusu.ate.theme.MyTheme
 
-class No3Fragment : Fragment() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enterTransition = MaterialSharedAxis(MaterialSharedAxis.X, true)
-        returnTransition = MaterialSharedAxis(MaterialSharedAxis.X, false)
-    }
+class Screen1Fragment : Fragment() {
 
     @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreateView(
@@ -44,22 +42,49 @@ class No3Fragment : Fragment() {
                         topBar = {
                             MyTopAppBar(
                                 title = {
-                                    Text(text = stringResource(id = R.string.screen_n, 3))
+                                    Text(text = stringResource(id = R.string.screen_n, 1))
                                 },
                                 hasBack = true,
                                 onBackPressed = {
-                                    findNavController().popBackStack()
+                                    requireActivity().finish()
                                 },
                             )
                         },
                     ) { innerPadding ->
-                        Text(
+                        Column(
                             modifier = Modifier
                                 .fillMaxSize()
                                 .padding(innerPadding)
                                 .padding(16.dp),
-                            text = stringResource(id = R.string.screen_n, 3),
-                        )
+                            verticalArrangement = spacedBy(16.dp),
+                        ) {
+                            Button(
+                                onClick = {
+                                    exitTransition = null
+                                    reenterTransition = null
+                                    findNavController().navigate(
+                                        Screen1FragmentDirections.actionScreen2(),
+                                    )
+                                },
+                            ) {
+                                Text(text = stringResource(id = R.string.open_screen_n, 2))
+                            }
+                            Button(
+                                onClick = {
+                                    exitTransition = MaterialSharedAxis(
+                                        MaterialSharedAxis.X, true,
+                                    )
+                                    reenterTransition = MaterialSharedAxis(
+                                        MaterialSharedAxis.X, false,
+                                    )
+                                    findNavController().navigate(
+                                        Screen1FragmentDirections.actionScreen3(),
+                                    )
+                                },
+                            ) {
+                                Text(text = stringResource(id = R.string.open_screen_n, 3))
+                            }
+                        }
                     }
                 }
             }
