+++ app/src/main/java/com/tfandkusu/ate/feature/jnta/Screen2Fragment.kt
@@ -20,7 +20,7 @@ import com.tfandkusu.ate.R
 import com.tfandkusu.ate.component.MyTopAppBar
 import com.tfandkusu.ate.theme.MyTheme
 
-class No2Fragment : Fragment() {
+class Screen2Fragment : Fragment() {
 
     @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreateView(
