+++ app/src/main/java/com/tfandkusu/ate/feature/jnta/Screen3Fragment.kt
@@ -4,11 +4,8 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.foundation.layout.Arrangement.Absolute.spacedBy
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -24,7 +21,12 @@ import com.tfandkusu.ate.R
 import com.tfandkusu.ate.component.MyTopAppBar
 import com.tfandkusu.ate.theme.MyTheme
 
-class No1Fragment : Fragment() {
+class Screen3Fragment : Fragment() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enterTransition = MaterialSharedAxis(MaterialSharedAxis.X, true)
+        returnTransition = MaterialSharedAxis(MaterialSharedAxis.X, false)
+    }
 
     @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreateView(
@@ -42,49 +44,22 @@ class No1Fragment : Fragment() {
                         topBar = {
                             MyTopAppBar(
                                 title = {
-                                    Text(text = stringResource(id = R.string.screen_n, 1))
+                                    Text(text = stringResource(id = R.string.screen_n, 3))
                                 },
                                 hasBack = true,
                                 onBackPressed = {
-                                    requireActivity().finish()
+                                    findNavController().popBackStack()
                                 },
                             )
                         },
                     ) { innerPadding ->
-                        Column(
+                        Text(
                             modifier = Modifier
                                 .fillMaxSize()
                                 .padding(innerPadding)
                                 .padding(16.dp),
-                            verticalArrangement = spacedBy(16.dp),
-                        ) {
-                            Button(
-                                onClick = {
-                                    exitTransition = null
-                                    reenterTransition = null
-                                    findNavController().navigate(
-                                        No1FragmentDirections.actionNo2(),
-                                    )
-                                },
-                            ) {
-                                Text(text = stringResource(id = R.string.open_screen_n, 2))
-                            }
-                            Button(
-                                onClick = {
-                                    exitTransition = MaterialSharedAxis(
-                                        MaterialSharedAxis.X, true,
-                                    )
-                                    reenterTransition = MaterialSharedAxis(
-                                        MaterialSharedAxis.X, false,
-                                    )
-                                    findNavController().navigate(
-                                        No1FragmentDirections.actionNo3(),
-                                    )
-                                },
-                            ) {
-                                Text(text = stringResource(id = R.string.open_screen_n, 3))
-                            }
-                        }
+                            text = stringResource(id = R.string.screen_n, 3),
+                        )
                     }
                 }
             }
